requirements.txt
python-telegram-bot[job-queue]~=20.7
httpx~=0.27
googletrans==4.0.0-rc1
APScheduler~=3.10
